+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#673* : Fixed issue with Stream processing filters that specify Create Time
+
 ## [v6.0-alpha.15]
 
 * Issue **#666** : Fixed the duplicate dictionary issue in processing filter migrations, made querying more efficient too